add fasta file pull

+++ generate_sample_table.sh
@@ -6,6 +6,7 @@ shopt -s extglob
 
 
 database_dir=0
+fasta_dir=0
 name="sample_table.csv"
 existing=0
 
@@ -13,32 +14,43 @@ HELP="""
 ASSUMES FASTQ FILES ARE NAMED AS <sample_name>_L00#_R{1,2}*.fastq(.gz)
 
 Flags:
-    -d  :  Path to directory containing sample fastq(.gz) files (Absolute paths preferred for consistency, but can use relative paths)
+    -d  :  Path to directory containing sample fastq(.gz) files (Absolute paths preferred for consistency, but can use relative paths) (incompatible with -f)
+    -f  :  Path to directory containing sample consensus FASTA files (only compatible with fasta_process) (incompatible with -d)
     -n  :  Name or file path for final sample table (with extension) (default: 'sample_table.csv') - will overwrite if file exists
-    -e  :  Name or file path for an existing sample table - will append to the end of the provided table
+    -e  :  Name or file path for an existing fastq(.gz) sample table - will append to the end of the provided table
 
 Select one of '-n' (new sample table) or '-e' (existing sample table). 
 If neither provided, a new sample table called 'sample_table.csv' will be created (or overwritten) by default. 
 """
 
-while getopts ":d:n:e:" option; do
+while getopts ":d:f:n:e:" option; do
     case "${option}" in
         d) database_dir=$OPTARG;;
+        f) fasta_dir=$OPTARG;;
         n) name=$OPTARG;;
         e) existing=$OPTARG;;
     esac
 done
 
-if [ $database_dir = 0 ] ; then
-    echo "You must specify a data directory containing fastq(.gz) reads."
+if [ $database_dir = 0 ] && [ $fasta_dir = 0 ]; then
+    echo "You must specify a data directory containing fastq(.gz) reads (-d) or consensus FASTA files (-f)."
+    echo "$HELP"
+    exit 1
+fi
+
+if [ ! $database_dir = 0 ] && [ ! $fasta_dir = 0 ]; then
+    echo "You can only select one directory of fastq(.gz) reads (-d) or consensus FASTA files (-f)."
     echo "$HELP"
     exit 1
 fi
 
 echo -e "Adding samples\n"
-if [ $existing = 0 ] ; then
+if [ $existing = 0 ] && [ ! $database_dir = 0 ]; then
 	echo -e "Creating new sample table: ${name}\n"
  	echo "sample,r1_path,r2_path" > ${name} && echo "sample,r1_path,r2_path"
+elif [ $existing = 0 ] && [ ! $fasta_dir = 0 ]; then
+        echo -e "Creating new sample table: ${name}\n"
+        echo "sample,fasta_path" > ${name} && echo "sample,fasta_path"
 else
 	filename=$(basename $existing)
 	if [ -f ""$existing"" ] ; then
@@ -68,6 +80,19 @@ if [ ! $database_dir = 0 ]; then
 	done
 	echo -e "\n"
 fi
+
+if [ ! $fasta_dir = 0 ]; then
+		samples_dir=()
+		samples_fail=()
+		for file in $fasta_dir/*.{fa,fasta}; do
+			sample=$(basename $file | cut -d_ -f 1 | cut -d. -f1)
+			if [ -f $file ]; then
+				echo ${sample},${file} >> ${name} && echo ${sample},${file}
+			fi
+		done
+		echo -e "\n"
+fi
+
 
 if [ ${#samples_fail[@]} -gt 0 ]; then 
 	echo -e "Samples that failed to be added to sample table:"
